requirements.txt
appdirs==1.4.4
argcomplete==3.6.0
asgiref==3.8.1
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.1
idna==3.10
importlib_resources==6.5.2
pillow==10.4.0
python-barcode==0.15.1
python-escpos==3.1
pywin32==308
PyYAML==6.0.2
qrcode==8.0
reportlab==4.3.1
requests==2.32.3
setuptools==76.0.0
six==1.17.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.3.0
